Created basic bookdown structure
bookdown needs to be all in one R project with an index file and chapter files numbered in the order you want them in the book.  I created the chapter files based on my recommended structure.  Please speak up of you don't like this structure!  We need to determine the structure early in the planning process or it will be a mess to change it later.